production-security-framework-backend-pro/backend/requirements.txt
```
fastapi==0.111.0
uvicorn==0.30.1
SQLAlchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
loguru==0.7.2
fastapi-limiter==0.1.6
redis==5.0.4
fastapi-cache2==0.2.1
fastapi-cache2[redis]==0.2.1
httpx==0.27.0
pytest==8.2.1
pytest-asyncio==0.23.6
pytest-alembic==0.9.1
asyncio-throttle==0.1.1
```